requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
DateTime==4.3
duration==1.1.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
subscription==0.1.1
uuid==1.30
Werkzeug==2.0.2
WTForms==2.3.3
zope.interface==5.4.0